Move `ServiceListItem` to `Icinga\Module\Icingadb\Widget\ItemList`
