+++ library/Icingadb/Widget/ItemList/ServiceListItem.php
@@ -2,7 +2,7 @@
 
 /* Icinga DB Web | (c) 2020 Icinga GmbH | GPLv2 */
 
-namespace Icinga\Module\Icingadb\Widget;
+namespace Icinga\Module\Icingadb\Widget\ItemList;
 
 use Icinga\Module\Icingadb\Common\ListItemCommonLayout;
 use ipl\Web\Widget\StateBall;
